Fix eukara's error-on-restart issue. Tweak downloads menu tooltips. Some minor cleanups.
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@5383 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/client/screen.h
@@ -323,4 +323,4 @@ fte_inline float M_LinearToSRGB(float x, float mag)
 
 void R_NetgraphInit(void);
 void R_NetGraph (void);
-void R_FrameTimeGraph (int frametime);
+void R_FrameTimeGraph (float frametime);
